Override inline props with props passed via animatedProps
