+++ packages/react-native-reanimated/__tests__/svg.test.tsx
@@ -2,8 +2,6 @@ import { fireEvent, render } from '@testing-library/react-native';
 import React from 'react';
 import { Pressable, Text } from 'react-native';
 import Animated, {
-  type CSSAnimationProperties,
-  type CSSTransitionProperties,
   useAnimatedProps,
   useAnimatedStyle,
   useSharedValue,
@@ -119,86 +117,5 @@ describe('Animating wrapper of SVG', () => {
     jest.advanceTimersByTime(600);
 
     expect(wrapper).toHaveAnimatedStyle({ opacity: 1 });
-  });
-});
-
-describe('Plain object animatedProps', () => {
-  function TestComponent({
-    animatedProps,
-  }: {
-    animatedProps: Record<string, unknown>;
-  }) {
-    return (
-      <Svg height="100" width="100">
-        <AnimatedCircle
-          testID="circle"
-          cx="50"
-          cy="50"
-          animatedProps={animatedProps}
-        />
-      </Svg>
-    );
-  }
-
-  test('forwards values to the underlying component on initial render', () => {
-    const { getByTestId } = render(
-      <TestComponent animatedProps={{ r: 20, fill: 'blue' }} />
-    );
-
-    const circle = getByTestId('circle');
-    expect(circle.props.r).toBe(20);
-    expect(circle.props.fill).toBe('blue');
-  });
-
-  test('updates the rendered values when animatedProps change', () => {
-    const { getByTestId, rerender } = render(
-      <TestComponent animatedProps={{ r: 20, fill: 'blue' }} />
-    );
-
-    rerender(<TestComponent animatedProps={{ r: 50, fill: 'red' }} />);
-    expect(getByTestId('circle').props.r).toBe(50);
-    expect(getByTestId('circle').props.fill).toBe('red');
-  });
-
-  test('does not forward CSS transition config keys', () => {
-    const transitionConfig: Required<CSSTransitionProperties> = {
-      transitionProperty: 'all',
-      transitionDuration: '500ms',
-      transitionTimingFunction: 'ease-in',
-      transitionDelay: '0ms',
-      transitionBehavior: 'normal',
-      transition: 'all 500ms ease-in 0ms',
-    };
-
-    const { getByTestId } = render(
-      <TestComponent animatedProps={transitionConfig} />
-    );
-
-    const circle = getByTestId('circle');
-    for (const key of Object.keys(transitionConfig)) {
-      expect(circle.props[key]).toBeUndefined();
-    }
-  });
-
-  test('does not forward CSS animation config keys', () => {
-    const animationConfig: Required<CSSAnimationProperties> = {
-      animationName: 'none',
-      animationDuration: '1s',
-      animationTimingFunction: 'linear',
-      animationDelay: '0ms',
-      animationIterationCount: 1,
-      animationDirection: 'normal',
-      animationFillMode: 'none',
-      animationPlayState: 'running',
-    };
-
-    const { getByTestId } = render(
-      <TestComponent animatedProps={animationConfig} />
-    );
-
-    const circle = getByTestId('circle');
-    for (const key of Object.keys(animationConfig)) {
-      expect(circle.props[key]).toBeUndefined();
-    }
   });
 });
